+++ src/components/LevelItem/index.tsx
@@ -1,6 +1,6 @@
-import { InfoIcon, CheckIcon } from "@chakra-ui/icons";
+import {  CheckIcon } from "@chakra-ui/icons";
 import { FaRegCircle } from "react-icons/fa";
-import { Center, Flex, Spacer, Stack } from "@chakra-ui/react";
+import { Center, Stack } from "@chakra-ui/react";
 import { Level, RoadmapItem } from "../../entity/RoadmapModel";
 import ReactGA from "react-ga4";
