+++ src/pages/CustomRoadmapPage.tsx
@@ -1,9 +1,9 @@
 import { useParams } from "react-router-dom";
 import MainLayout from "../components/layouts/MainLayout";
-import Roadmap from "../components/Roadmap/Roadmap";
+import Roadmap from "../components/Roadmap";
 import E404Page from "./E404Page";
 import useDocumentTitle from "../components/useDocumentTitle";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import axios from "axios";
 import Cookies from "universal-cookie";
 import { RoadmapModel } from "../entity/RoadmapModel";
