+++ src/pages/GuidePage.tsx
@@ -1,14 +1,8 @@
 import { useParams } from "react-router-dom";
 import MainLayout from "../components/layouts/MainLayout";
-import Roadmap from "../components/Roadmap/Roadmap";
-import E404Page from "./E404Page";
 import useDocumentTitle from "../components/useDocumentTitle";
-import { useEffect, useRef, useState } from "react";
-import axios from "axios";
+import { useEffect } from "react";
 import Cookies from "universal-cookie";
-import { RoadmapModel } from "../entity/RoadmapModel";
-import { Grid } from "react-loader-spinner";
-
 
 const cookies = new Cookies();
