+++ src/pages/RoadmapPage.tsx
@@ -1,6 +1,6 @@
-import { useLocation, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import MainLayout from "../components/layouts/MainLayout";
-import Roadmap from "../components/Roadmap/Roadmap";
+import Roadmap from "../components/Roadmap";
 import { data as frontendData } from "../roadmaps/frontend";
 import { data as reactData } from "../roadmaps/react";
 import { data as backendData } from "../roadmaps/backend";
@@ -8,7 +8,7 @@ import { data as devopsData } from "../roadmaps/devops";
 import { data as communityData } from "../roadmaps/community";
 import E404Page from "./E404Page";
 import useDocumentTitle from "../components/useDocumentTitle";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { dataEngineeringData } from "../roadmaps/dataEngineering";
 
 export default function RoadmapPage() {
